domain: add OBSManager and simplify SlpFileWriter
