+++ app/domain/SlpFileWriter.js
@@ -29,7 +29,6 @@ import net from 'net';
 import _ from 'lodash';
 import path from 'path';
 import moment from 'moment';
-import OBSWebSocket from 'obs-websocket-js';
 import {
   Ports,
   SlpFile,
@@ -37,24 +36,18 @@ import {
   SlpStream,
   SlpStreamEvent,
 } from '@slippi/slippi-js';
+import OBSManager from './OBSManager';
 
 export default class SlpFileWriter extends EventEmitter {
   constructor(settings) {
     super();
     this.folderPath = settings.folderPath;
     this.onFileStateChange = settings.onFileStateChange;
-    this.obsSourceName = settings.obsSourceName;
-    this.obsIP = settings.obsIP;
-    this.obsPassword = settings.obsPassword;
     this.id = settings.id;
     this.consoleNick = settings.consoleNick;
     this.currentFile = this.getClearedCurrentFile();
-    this.obs = new OBSWebSocket();
+    this.obs = new OBSManager(settings);
     this.slpStream = new SlpStream();
-    this.statusOutput = {
-      status: false,
-      timeout: null,
-    };
     this.isRelaying = settings.isRelaying;
     this.clients = [];
     this.manageRelay();
@@ -120,46 +113,15 @@ export default class SlpFileWriter extends EventEmitter {
 
   updateSettings(settings) {
     this.folderPath = settings.targetFolder;
-    this.obsIP = settings.obsIP;
-    this.obsSourceName = settings.obsSourceName;
-    this.obsPassword = settings.obsPassword;
     this.id = settings.id;
     this.isRelaying = settings.isRelaying;
     this.consoleNick = settings.consoleNick || this.consoleNick;
+    this.obs.updateSettings(settings);
     this.manageRelay();
   }
 
-  getSceneSources = async () => {
-    // eslint-disable-line
-    const res = await this.obs.send('GetSceneList');
-    const scenes = res.scenes || [];
-    const pairs = _.flatMap(scenes, scene => {
-      const sources = scene.sources || [];
-      return _.map(sources, source => ({
-        scene: scene.name,
-        source: source.name,
-      }));
-    });
-    this.obsPairs = _.filter(pairs, pair => pair.source === this.obsSourceName);
-  };
-
-  async connectOBS() {
-    if (this.obsIP && this.obsSourceName) {
-      // if you send a password when authentication is disabled, OBS will still connect
-      await this.obs.connect({
-        address: this.obsIP,
-        password: this.obsPassword,
-      });
-      await this.obs.on(
-        'SceneItemAdded',
-        async () => this.getSceneSources()
-      ); // eslint-disable-line
-      await this.obs.on(
-        'SceneItemRemoved',
-        async () => this.getSceneSources()
-      ); // eslint-disable-line
-      await this.getSceneSources();
-    }
+  connectOBS() {
+    this.obs.connect();
   }
 
   disconnectOBS() {
@@ -168,49 +130,6 @@ export default class SlpFileWriter extends EventEmitter {
     }
 
     this.obs.disconnect();
-  }
-
-  setStatus(value) {
-    this.statusOutput.status = value;
-    // console.log(`Status changed: ${value}`);
-    _.forEach(this.obsPairs, pair => {
-      this.obs.send('SetSceneItemProperties', {
-        'scene-name': pair.scene,
-        item: this.obsSourceName,
-        visible: value,
-      });
-    });
-  }
-
-  handleStatusOutput(timeoutLength = 200) {
-    const setTimer = () => {
-      if (this.statusOutput.timeout) {
-        // If we have a timeout, clear it
-        clearTimeout(this.statusOutput.timeout);
-      }
-
-      this.statusOutput.timeout = setTimeout(() => {
-        // If we timeout, set and set status
-        this.setStatus(false);
-      }, timeoutLength);
-    };
-
-    if (this.currentFile.metadata.lastFrame < -60) {
-      // Only show the source in the later portion of the game loading stage
-      return;
-    }
-
-    if (this.statusOutput.status) {
-      // If game is currently active, reset the timer
-      setTimer();
-      return;
-    }
-
-    // Here we did not have a game going, so let's indicate we do now
-    this.setStatus(true);
-
-    // Set timer
-    setTimer();
   }
 
   handleData(newData) {
@@ -254,7 +173,7 @@ export default class SlpFileWriter extends EventEmitter {
         break;
       default:
         this.writeCommand(payload);
-        this.handleStatusOutput();
+        this.obs.handleStatusOutput();
         break;
       }
     });
@@ -294,7 +213,7 @@ export default class SlpFileWriter extends EventEmitter {
         break;
       case Command.GAME_END:
         if (payload.gameEndMethod !== 7) {
-          this.handleStatusOutput(700);
+          this.obs.handleStatusOutput(700);
         }
         break;
       default:
